authorization

+++ app/Http/Controllers/TodoController.php
@@ -7,11 +7,6 @@ use Illuminate\Http\Request;
 
 class TodoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $todoes = Todo::all();
@@ -19,11 +14,6 @@ class TodoController extends Controller
         return view('todoes.index', $data);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         $todo = new Todo();
@@ -31,12 +21,6 @@ class TodoController extends Controller
         return view('todoes.create', $data);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $this->validate($request, [
@@ -52,39 +36,22 @@ class TodoController extends Controller
         return redirect(route('todoes.index'));
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Todo  $todo
-     * @return \Illuminate\Http\Response
-     */
     public function show(Todo $todo)
     {
         $data = ['todo' => $todo];
         return view('todoes.show', $data);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Todo  $todo
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Todo $todo)
     {
+        $this->authorize($todo);
         $data = ['todo' => $todo];
         return view('todoes.edit', $data);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Todo  $todo
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, Todo $todo)
     {
+        $this->authorize($todo);
         $this->validate($request, [
             'title' => 'required|max:255',
             'body'=> 'required|max:255'
@@ -95,14 +62,9 @@ class TodoController extends Controller
         return redirect(route('todoes.show', $todo));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Todo  $todo
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Todo $todo)
     {
+        $this->authorize($todo);
         $todo->delete();
         return redirect(route('todoes.index'));
     }
